ggvenn: Draw Venn Diagram by 'ggplot2'

An easy-to-use way to draw pretty Venn diagrams using 'ggplot2'. This package provides functions to create Venn diagrams with customizable colors, labels, and styling options.

See also

Author

Maintainer: Linlin Yan yanlinlin82@gmail.com (ORCID)